Organize similar classes into packages
